Add files via upload
In this project, Cookie Clicker was an amazing app to create and a good first experience in coding. The only thing I found difficult in this was customizing the app like centering the buttons or adding a picture of a cookie. But despite that, the rest of the coding was a fun experience for me.